feat: restructure types dir
