+++ src/types/category/methods.ts
@@ -1,0 +1,7 @@
+import { Budget } from "../budget/types";
+import { DateRange } from "../utils/types";
+import { Category, Period } from "./types";
+
+export function calcPeriods(budget: Budget, category: Category): Period[] {
+  return [];
+}
