+++ src/types/utils/types.ts
@@ -1,0 +1,1 @@
+export type DateRange = { begin: Date; end: Date };
